Improves NR7 edit overview workflow and validation
Enhances the NR7 edit overview by adding workflow actions within the preview section when a document is locked, ensuring a smoother user experience.

Also, prevents publishing when there is no NR7 document to publish and no NR7 document was modified but indicator data was, create a draft to allow publishing.

Renames the storage key to properly reflect the scope.